requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.9.4=pypi_0
aiosignal=1.3.1=pypi_0
alpaca-eval=0.1.2=pypi_0
altair=4.2.2=pypi_0
annotated-types=0.6.0=py39h06a4308_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli-python=1.0.9=py39h6a678d5_7
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2024.2.2=py39h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.2.1=pypi_0
cpm-kernels=1.0.11=pypi_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
datasets=2.10.1=pypi_0
deepspeed=0.8.3=pypi_0
dill=0.3.6=pypi_0
distro=1.8.0=py39h06a4308_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.2.0=py39h06a4308_0
fairscale=0.4.13=pypi_0
fastapi=0.95.0=pypi_0
ffmpy=0.3.0=pypi_0
filelock=3.13.1=py39h06a4308_0
fire=0.5.0=pypi_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fschat=0.2.28=pypi_0
fsspec=2024.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
google-auth=2.17.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gpustat=1.1=pypi_0
gradio=3.23.0=pypi_0
grpcio=1.53.0=pypi_0
h11=0.14.0=py39h06a4308_0
hjson=3.1.0=pypi_0
httpcore=0.16.3=pypi_0
httpx=0.23.3=pypi_0
huggingface-hub=0.21.4=pypi_0
icetk=0.0.4=pypi_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py39h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.5=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libwebp-base=1.3.2=h5eee18b_0
linkify-it-py=2.0.0=pypi_0
loralib=0.1.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markdown2=2.4.10=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.4=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py39h06a4308_0
nh3=0.2.14=pypi_0
ninja=1.11.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-ml-py=11.525.84=pypi_0
oauthlib=3.2.2=pypi_0
openai=1.9.0=py39h06a4308_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
orjson=3.8.8=pypi_0
packaging=24.0=pypi_0
pandas=2.2.1=py39h6a678d5_0
pandas-stubs=2.1.4.231227=py39h06a4308_0
peft=0.7.1=pypi_0
pillow=10.2.0=py39h5eee18b_0
pip=24.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=5.26.1=pypi_0
psutil=5.9.8=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydantic=1.10.7=pypi_0
pydantic-core=2.14.6=py39hb02cf49_0
pydub=0.25.1=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.2=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-multipart=0.0.6=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.0.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.22=pypi_0
requests=2.31.0=py39h06a4308_1
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3986=1.5.0=pypi_0
rich=13.5.2=pypi_0
rouge=1.0.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-fuzzy=0.4.2=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=py39hdb19cb5_0
sentry-sdk=1.45.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39h06a4308_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.0=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
starlette=0.26.1=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
textbrewer=0.2.1.post1=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.5.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.2=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.2=cpu_py39h83e0c9b_0
tqdm=4.65.0=py39hb070fc8_0
transformers=4.36.2=pypi_0
transformers-stream-generator=0.0.5=pypi_0
types-pytz=2022.4.0.0=py39h06a4308_1
typing-extensions=4.9.0=py39h06a4308_1
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024a=h04d1e81_0
uc-micro-py=1.0.1=pypi_0
urllib3=2.1.0=py39h06a4308_1
uvicorn=0.21.1=pypi_0
wandb=0.16.6=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.6=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.41.2=py39h06a4308_0
xxhash=3.2.0=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
